backend/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
Werkzeug==1.0.0
Flask-Cors==3.0.8
requests==2.23.0